+++ src/pages/heylynnShin/Login/Login.scss
@@ -6,6 +6,7 @@ main.login {
   align-items: center;
   background: #f1f3f5;
   min-height: 100vh;
+
   .inner {
     width: 350px;
     height: 450px;
@@ -13,11 +14,13 @@ main.login {
     box-sizing: border-box;
     background: #fff;
     border: 1px solid #dbdbdb;
+
     .header {
       display: flex;
       justify-content: center;
       align-items: center;
       margin: 30px 0;
+
       h1 {
         display: flex;
         font-family: 'Lobster Two', cursive;
@@ -25,8 +28,10 @@ main.login {
       }
     }
   }
+
   .form {
     margin-bottom: 30px;
+
     .inputWrapper {
       input {
         width: 100%;
@@ -46,6 +51,7 @@ main.login {
       width: 100%;
       height: 40px;
       margin-bottom: 70px;
+
       .activeBtn {
         width: 100%;
         height: 100%;
@@ -53,6 +59,7 @@ main.login {
         border: 1px solid #0095f6;
         border-radius: 4px;
         cursor: pointer;
+
         span {
           font-size: 18px;
           line-height: 1.4;
@@ -60,12 +67,14 @@ main.login {
           color: #fff;
         }
       }
+
       .btn {
         width: 100%;
         height: 100%;
         background-color: #bfdffd;
         border: 1px solid #bfdffd;
         border-radius: 4px;
+
         span {
           font-size: 18px;
           line-height: 1.4;
@@ -75,12 +84,15 @@ main.login {
       }
     }
   }
+
   .passwordForget {
     display: flex;
     justify-content: center;
     margin-top: 40px;
+
     .pageMove {
       display: block;
+
       span {
         font-size: 15px;
       }
