+++ src/pages/heylynnShin/Main/CommentInput.scss
@@ -1,0 +1,8 @@
+ul li {
+  font-size: 14px;
+  padding-left: 5px;
+  padding-bottom: 3px;
+  a {
+    font-weight: 600;
+  }
+}
