+++ src/pages/heylynnShin/Main/Feed.scss
@@ -1,0 +1,158 @@
+.mainContainer {
+  display: flex;
+  overflow: hidden;
+  justify-content: right;
+  padding-right: 30px;
+
+  .feeds {
+    width: 80%;
+    padding-top: 20px;
+    // border: 1px solid lightgray;
+    display: flex;
+    flex-direction: column;
+
+    .feed {
+      border: 1px solid lightgray;
+      display: flex;
+      flex-direction: column;
+      margin-left: auto;
+      background-color: white;
+
+      .profileContainer {
+        height: 45px;
+        display: flex;
+        padding: 4px 0;
+
+        .feedProfile {
+          width: 80px;
+          display: flex;
+          align-items: center;
+          justify-content: space-around;
+          padding: 5px;
+
+          .imageAndIdContainer {
+            padding-left: 15px;
+            display: flex;
+            align-items: center;
+            justify-content: space-around;
+
+            .imageConatiner {
+              border-radius: 70%;
+              overflow: hidden;
+              width: 30px;
+              height: 30px;
+              display: flex;
+
+              .image {
+                width: 100%;
+                height: 100%;
+                object-fit: cover;
+              }
+            }
+
+            .id {
+              font-size: 15px;
+              padding-left: 5px;
+              font-weight: 600;
+            }
+          }
+        }
+      }
+
+      .feedImageContainer {
+        .feedImage {
+          width: 100%;
+          object-fit: cover;
+        }
+      }
+
+      .belowFeedImage {
+        padding-left: 5px;
+
+        .feedIconContainer {
+          display: flex;
+          justify-content: space-between;
+
+          .iconContainer {
+            display: flex;
+            width: 60px;
+            justify-content: space-between;
+          }
+
+          .bookmarkIconContainer {
+            padding-right: 10px;
+          }
+        }
+
+        .feedText {
+          font-size: 13px;
+
+          .postContainer {
+            padding-top: 10px;
+            display: flex;
+            align-items: center;
+
+            .id {
+              font-weight: bold;
+            }
+
+            .post {
+              padding-left: 5px;
+            }
+
+            .more {
+              color: gray;
+              padding-left: 5px;
+            }
+          }
+
+          .writtenCommentContainer {
+            padding-top: 5px;
+            display: flex;
+            align-items: center;
+
+            .id {
+              font-weight: bold;
+            }
+            .writtenComment {
+              padding-left: 5px;
+            }
+          }
+
+          .timeConatiner {
+            .time {
+              margin: 0;
+              padding-bottom: 5px;
+              color: gray;
+            }
+          }
+        }
+      }
+
+      .commentContainer {
+        display: flex;
+        justify-content: space-between;
+        padding-top: 10px;
+        border-top: 1px solid lightgray;
+
+        .inputComment {
+          font-size: 12px;
+          border: none;
+          width: 95%;
+        }
+
+        .inputComment:focus {
+          outline: none;
+        }
+
+        .commentBtn {
+          white-space: nowrap;
+          border: none;
+          background-color: transparent;
+          color: lightskyblue;
+          cursor: pointer;
+        }
+      }
+    }
+  }
+}
