+++ src/pages/heylynnShin/Main/Liked.scss
@@ -1,0 +1,34 @@
+.likeContainer {
+  hegith: 30px;
+  display: flex;
+  align-items: center;
+  padding-top: 5px;
+
+  .likedIdContainer {
+    display: flex;
+    align-items: center;
+
+    .imageContainer {
+      border-radius: 70%;
+      overflow: hidden;
+      width: 20px;
+      height: 20px;
+
+      .image {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+      }
+    }
+
+    .idContainer {
+      padding-left: 5px;
+      .id {
+        font-weight: bold;
+      }
+    }
+  }
+  .like a {
+    font-weight: bold;
+  }
+}
