+++ src/pages/heylynnShin/Main/Main.scss
@@ -1,212 +1,14 @@
 @import url('https://fonts.googleapis.com/css2?family=Lobster+Two&family=Noto+Serif+Armenian:wght@500&display=swap');
 
-.main {
+.westagram {
   background-color: #fafafa;
-  nav {
-    width: 100%;
-    height: 50px;
-    display: flex;
-    align-items: center;
-    justify-content: space-around;
-    border-bottom: 1px solid lightgray;
-    background-color: white;
-    .logo {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      .logoImage {
-        width: 30px;
-        padding-right: 10px;
-      }
-      .logoText {
-        font-size: 20px;
-        border-left: 1px solid;
-        padding-left: 10px;
-        font-family: 'Lobster Two', cursive;
-      }
-    }
-    .searchBar {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      border: 1px solid lightgray;
-      .searchBarInput {
-        width: 150px;
-        height: 20px;
-        border: none;
-        font-size: 12px;
-        background-color: #fafafa;
-        text-align: center;
-      }
-    }
-    .navIcon {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      .compass,
-      .heart,
-      .mypage {
-        width: 30px;
-        padding: 5px;
-        margin-top: -5px;
-      }
-    }
-  }
 
-  main {
+  .main {
+    padding-top: 30px;
     display: flex;
-    justify-content: space-around;
-    align-items: center;
+    justify-content: center;
     overflow: hidden;
-    .feeds {
-      max-width: 800px;
-      width: 70%;
-      padding-top: 20px;
-      border: 1px solid lightgray;
-      margin: 0px 20px 0px 50px;
-      display: flex;
-      .feed {
-        height: 100%;
-        border: 1px solid lightgray;
-        display: flex;
-        flex-direction: column;
-        margin-left: auto;
-        background-color: white;
-        .profileContainer {
-          height: 35px;
-          display: flex;
-          padding: 4px 0;
-          .feedProfile {
-            width: 70px;
-            display: flex;
-            align-items: center;
-            justify-content: space-around;
-            padding: 5px;
-            .imageAndIdContainer {
-              padding-left: 20px;
-              display: flex;
-              align-items: center;
-              justify-content: space-around;
-              .imageConatiner {
-                border-radius: 70%;
-                overflow: hidden;
-                width: 25px;
-                height: 25px;
-                display: flex;
-                .image {
-                  width: 100%;
-                  height: 100%;
-                  object-fit: cover;
-                }
-              }
-              .id {
-                font-size: 13px;
-                padding-left: 10px;
-              }
-            }
-          }
-        }
-        .feedImageContainer {
-          .feedImage {
-            width: 100%;
-            object-fit: cover;
-          }
-        }
-        .belowFeedImage {
-          padding-left: 5px;
-          .feedIconContainer {
-            display: flex;
-            justify-content: space-between;
-            .iconContainer {
-              display: flex;
-              width: 60px;
-              justify-content: space-between;
-            }
-            .bookmarkIconContainer {
-              padding-right: 10px;
-            }
-          }
-          .feedText {
-            font-size: 12px;
-            .likeContainer {
-              .like {
-                margin-top: 5px;
-              }
-            }
-            .postContainer {
-              .post {
-                margin-top: 10px;
-                margin-bottom: 0;
-              }
-            }
-            .writtenCommentContainer {
-              .writtenComment {
-                margin: 0;
-              }
-            }
-            .timeConatiner {
-              .time {
-                margin: 0;
-                padding-bottom: 5px;
-                color: gray;
-              }
-            }
-          }
-        }
-        .commentContainer {
-          display: flex;
-          justify-content: space-between;
-          padding-top: 10px;
-          border: 1px solid lightgray;
-          .inputComment {
-            font-size: 12px;
-            border: none;
-          }
-          .submitCommentActive {
-            color: #0095f6;
-            background-color: transparent;
-            border: none;
-          }
-          .submitCommentInactive {
-            color: lightskyblue;
-            background-color: transparent;
-            border: none;
-          }
-        }
-      }
-    }
-  }
-  .feedsRightContainer {
-    display: flex;
-    width: 30%;
-    height: 100%;
-    .feedsRight {
-      width: 100%;
-      flex-direction: column;
-      .profile {
-        height: 50px;
-        border: 1px solid lightgray;
-      }
-      .story {
-        height: 150px;
-        border: 1px solid lightgray;
-      }
-      .recommendation {
-        height: 130px;
-        border: 1px solid lightgray;
-      }
-      .footer {
-        height: 60px;
-        border: 1px solid lightgray;
-        font-size: 10px;
-        .footerMenu li {
-          display: inline;
-          a {
-            text-decoration: none;
-            color: gray;
-          }
-        }
-      }
-    }
+    height: 100vh;
+    width: 100vw;
   }
 }
