+++ src/pages/heylynnShin/Main/Nav.scss
@@ -1,0 +1,59 @@
+@import url('https://fonts.googleapis.com/css2?family=Lobster+Two&family=Noto+Serif+Armenian:wght@500&display=swap');
+
+.nav {
+  width: 100%;
+  height: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  border-bottom: 1px solid lightgray;
+  background-color: white;
+
+  .logo {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    .logoImage {
+      width: 30px;
+      padding-right: 10px;
+    }
+
+    .logoText {
+      font-size: 20px;
+      border-left: 1px solid;
+      padding-left: 10px;
+      font-family: 'Lobster Two', cursive;
+    }
+  }
+
+  .searchBar {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: 1px solid lightgray;
+
+    .searchBarInput {
+      width: 150px;
+      height: 20px;
+      border: none;
+      font-size: 12px;
+      background-color: #fafafa;
+      text-align: center;
+    }
+  }
+
+  .navIcon {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    .compass,
+    .heart,
+    .mypage {
+      width: 30px;
+      padding: 5px;
+      margin-top: -5px;
+    }
+  }
+}
