+++ src/pages/heylynnShin/Main/Recommendation.scss
@@ -1,0 +1,60 @@
+.recommendProfile {
+  padding-left: 5px;
+  height: 40px;
+  display: flex;
+  align-items: center;
+  font-size: 13px;
+
+  .profileContainer {
+    display: flex;
+    align-items: center;
+    width: 100%;
+
+    .imageConatiner {
+      border-radius: 70%;
+      overflow: hidden;
+      width: 30px;
+      height: 30px;
+
+      .image {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+      }
+    }
+
+    .idContainer {
+      padding-left: 10px;
+      max-width: 70%;
+      display: flex;
+      justify-content: space-between;
+      width: 100%;
+
+      .idInfo {
+        width: 100%;
+
+        .id {
+          font-weight: bold;
+        }
+
+        .followers {
+          width: 80%;
+          color: gray;
+          overflow: hidden;
+          text-overflow: ellipsis;
+          white-space: nowrap;
+        }
+      }
+      .followBtn {
+        white-space: nowrap;
+        border: none;
+        background-color: transparent;
+        color: #0995f6;
+        padding: 0;
+        font-size: 13px;
+        font-weight: 600;
+        cursor: pointer;
+      }
+    }
+  }
+}
