+++ src/pages/heylynnShin/Main/Story.scss
@@ -1,0 +1,38 @@
+.storyProfile {
+  padding-left: 5px;
+  height: 40px;
+  display: flex;
+  align-items: center;
+
+  .profileContainer {
+    display: flex;
+    align-items: center;
+
+    .imageConatiner {
+      border-radius: 70%;
+      border: 2px solid #d300c5;
+      overflow: hidden;
+      width: 30px;
+      height: 30px;
+
+      .image {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+      }
+    }
+
+    .idContainer {
+      padding-left: 10px;
+      font-size: 13px;
+
+      .id {
+        font-weight: bold;
+      }
+
+      .time {
+        color: gray;
+      }
+    }
+  }
+}
