+++ src/pages/heylynnShin/Main/commentInput.js
@@ -1,8 +1,13 @@
+import './CommentInput.scss';
+
 const CommentInput = ({ comment }) => {
   return (
     <ul>
       {comment.map(comment => (
-        <li key={comment}>heylynn: {comment}</li>
+        <li key={comment}>
+          <a>heylynn: </a>
+          {comment}
+        </li>
       ))}
     </ul>
   );
